Better activation of existing instances
This detects if the SHM is still present, marked as active, but the process is no longer attached.

It will now only activate an existing instance if the process exists, and has the same execuatble name.

This is particularly useful when running from VSCode, if you click 'stop' in the debugger instead of exiting cleanly.