Regression: fix restarting OpenKneeboard without restarting the games
When adding configurable VR opacity, I made the kneeboard renderers cache some of the textures/mutexes etc. This keeps them alive even when the main OpenKneeboard app is dead.

As the mutex can only be relocked with the same key it was unlocked with, this means the texture *always* times out.

We could start from the previous key, but this would mean that the writer would have to read the SHM first. It's simplest to just make this detectable, throw away everything, and start again.