FIx event lifetime bugs
Initially thought this was concurrency bugs, but not really: either the sender or receiver can be destructed first, and this must be handled cleanly.

Additionally, EventReceivers *must* be disconnected before subclass destruction cleans this up at all. Check this on teardown, and add everywhere - otherwise it's possible for other threads to invoke the handlers while they're in the process of being torn down.