Fix OculusRoomTiny_DX11
OpenKneeboard is nuking the ID3D11DeviceContext while it's drawing. This is fine for most games (e.g. DCS, Falcon BMS, even OculusWorldDemo), but not for this one.

Fixing with `vrperfkit::StoreD3D11State` and `vrperfkit::RestoreD3D11State` - thanks @fholger :)

I've truncated vrperfkit's license file to exclude the third-party
licenses, as the relevant software is not included in this import

fixes #185