Document RELOAD_CURRENT_TAB in API messages
Added a note that this message is new as of v1.5 (assumption: it will be included in the next release with that number).

Also, noted that users should be careful in using this capability since the side effects could wipe out a lot of work.  There is no undo for the reload.