Feature/remote control reload current tab (#333)
Proposed addition of RELOAD_CURRENT_TAB as a user action.

These changes really just extend what's already there to make the
ReloabTab available as a user action. That allows it to be mapped in a
way so as not to require mouse interaction for invocation. This can be
quite handy in games where the time and concentration required to
provide the mouse inputs is impractical.

As requested, this does not include a way to bind the user action to a
button.

Also, this addition to UserActionHandler() for tab reload follows the
pattern for other actions made here and returns a std::uniqueptr. There
was talk about using sharedptr instead but since that would require
changing the function signature, I decided to leave that alone for now.