🦺 make tab construction async, add custom coroutine type
Well, this is a big pile of yak shaving.

Original goal was to prevent crashes when switching profiles if web
dashboard tabs are present

That led to finishing executing one user action before starting the next

That led to realizing that `Init()` must complete before `Create`
returns

That led to `Create()` needing to be async

That led to a massive refactor to `concurrent::task<T>`

That led to finding out that the `concurrent::task<T>` will sometimes
switch threads at `co_return`, even if you're on the original thread.

That led to trying `wil::com_task<>` which sounds like exactly what I
want

That has led to failures about missing and duplicate definitions from
windows headers. I spend some time debugging them, including removing
the obsolete references to `appmodel` (left over from when OpenKneeboard
used MSIX)

That led to writing a custom coroutine type which always resumes in the
thread it started in.