Make `make_stoppable()` private
This is only being used by resume_on_signal and resume_after - it should stay that way.

The WinRT cancel thing is great when you're actually dealing with WinRT, but the fact it's exception-based isn't.

Fixing that would require making my own resume_on_signal and resume_after based on threadpoolwait/threadpooltimeout, but that's not worth it