🔥 Remove legacy oculus (OVR) + D3D12 support
I'm not aware of this *ever* being used except for developer testing; it has major issues, and doesn't seem worth the maintenance time.

I may bring it back if there are pre-existing games that use it, are relevant to OpenKneeboard's intended use cases, and seem worth the maintenance burden.

I will not bring it back for new titles as OpenXR has been recommended over the legacy oculus API for several years now.